Refactorisation
Create Mover3D
Simplify Mover

+++ src/nc/Attractor.java
@@ -13,7 +13,7 @@ public class Attractor {
 	this.p = p;
 	location = new PVector(p.width/2, p.height/2);
 	mass = 20;
-	G = (float)0.4;
+	G = 0.4f;
 	}
 	
 	void display(){
@@ -25,7 +25,7 @@ public class Attractor {
 	PVector attract(Mover m){
 		PVector force = PVector.sub(location, m.location);
 		float distance = force.mag();
-		distance = p.constrain(distance,(float)5,(float)25);
+		distance = p.constrain(distance,5f,25f);
 		
 		force.normalize();
 		float strenght = (G * mass * m.mass)/(distance*distance);
